pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "word2cin"
authors = [{name = "Lôo Giān-tîng", email = "perf.patrick.lu@gmail.com"}]
version = "0.0.1"
description = "Generate cin input method file from Taigi dictionary sources"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "pyyaml",
    "openpyxl",
]

[tool.pytest.ini_options]
pythonpath = [
    ".",
    "src",
    "tests",
]
